+++ extensions/cursor-directory/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Cursor Directory Changelog
 
+## [Refactors and Improvements] - 2024-09-10
+
+- Fetch prompts from API instead of github repo
+- Fix display of prompt details
+
 ## [Initial Version] - 2024-09-10